Attempting to play opus on the fly
This doesn't work yet.

+++ source/main.c
@@ -171,8 +171,8 @@ int playWav(const char *wav)
 	u8		bitness;
 	u32		byterate; // TODO: Not used.
 	u32		blockalign;
-	u32*	buffer1 = NULL;
-	u32*	buffer2 = NULL;
+	u16*	buffer1 = NULL;
+	u16*	buffer2 = NULL;
 	ndspWaveBuf waveBuf[2];
 	bool playing = true;
 	bool lastbuf = false;
@@ -183,6 +183,7 @@ int playWav(const char *wav)
 		goto out;
 	}
 
+	// TODO: Check if this is required.
 	ndspSetOutputMode(NDSP_OUTPUT_STEREO);
 
 	if(file == NULL)
@@ -255,8 +256,8 @@ int playWav(const char *wav)
 	ndspChnSetFormat(CHANNEL, bitness);
 	memset(waveBuf, 0, sizeof(waveBuf));
 
-	buffer1 = (u32*) linearAlloc(BUFFER_SIZE);
-	buffer2 = (u32*) linearAlloc(BUFFER_SIZE);
+	buffer1 = (u16*) linearAlloc(BUFFER_SIZE);
+	buffer2 = (u16*) linearAlloc(BUFFER_SIZE);
 
 	fread(buffer1, 1, BUFFER_SIZE, file);
 	fread(buffer2, 1, BUFFER_SIZE, file);
